Struct efi::BootServices

pub struct BootServices {
    // some fields omitted
}

UEFI boot services.

A UEFI system has two modes: boot mode and runtime mode. In boot mode, the firmware is in control of the system; in return, the client has access to boot services provided by the firmware. To load an OS, the client must transition to runtime mode by calling exit_boot_services(). As its name implies, this call disables boot services and hands control of the system to the client.

Methods

impl BootServices

fn stdout(&self) -> &SimpleTextOutput

Returns a handle to the console output.

fn stderr(&self) -> &SimpleTextOutput

Returns a handle to the console standard error.

Note that in most implementations, any text written to this handle will not be displayed by default.

fn boxed<T>(&self, value: T) -> EfiBox<T>

Places an object on the UEFI heap.

Panics if the allocation fails.

unsafe fn allocate(&self, size: usize) -> *mut u8

Allocates a block of memory using the UEFI allocator. The memory is of type EfiLoaderData.

Panics if the allocation fails.

unsafe fn deallocate(&self, buffer: *mut u8)

Deallocates a block of memory provided by allocate().

fn locate_protocol<P: Protocol>(&self) -> Option<&P>

Returns the first protocol instance that matches the given protocol.

FIXME: This shouldn't return a &-reference, but a smart pointer that closes the protocol on drop.

Example

let output = boot_services.locate_protocol::<SimpleTextOutput>()
    .expect("could not find text output protocol");
write!(output, "Hello, world!\r\n").unwrap();

fn memory_map(&self) -> (MemoryMap, MapKey)

Retrieves a copy of the UEFI memory map.

fn exit_boot_services(self, key: MapKey) -> Result<(), (Status, BootServices)>

Terminate boot services.

unsafe fn get_instance() -> Option<BootServices>

Retrieves a copy of the boot services table, if present.

This method is useful for writing panic handlers, since they don't have direct access to the system table.

Safety

The caller must not use the instance after .exit_boot_services() is called.

Examples

#[lang = "panic_fmt"]
fn panic_fmt(args: core::fmt::Arguments, file: &str, line: u32) -> ! {
    // Log the panic to the console
    unsafe {
        if let Some(bs) = BootServices::get_instance() {
            let _ = write!(bs.stderr(), "PANIC: {} {} {}\r\n",
                           args, file, line);
        }
    }
    loop {}
}