Struct efi::GraphicsOutput

pub struct GraphicsOutput(_);

Methods

impl GraphicsOutput

fn linear_frame_buffer(&self) -> (PhysicalAddress, usize)

Returns the physical address and size of the linear frame buffer.

fn current_mode(&self) -> ModeNumber

Returns the index of the current mode.

fn current_mode_info(&self) -> &ModeInformation

Returns information pertaining to the current mode.

fn max_mode(&self) -> ModeNumber

Returns the number of modes supported by this device. All mode numbers are in the range [0, max_mode).

fn query_mode(&self, mode: ModeNumber) -> EfiResult<EfiBox<ModeInformation>>

Queries information on the specified mode.

fn set_mode(&self, mode: ModeNumber) -> EfiResult<()>

Sets the mode of this device.

fn fill(&self, x: usize, y: usize, w: usize, h: usize, pixel: BltPixel) -> EfiResult<()>

Fills the rectangle with a single color.

fn copy_buffer_to_video(&self, sx: usize, sy: usize, x: usize, y: usize, w: usize, h: usize, buffer: &[BltPixel], row_len: usize) -> EfiResult<()>

Copies a pixel buffer to the screen.