Struct efi::MemoryMap

pub struct MemoryMap {
    // some fields omitted
}

Represents a UEFI memory map.

Internally, this is an array of MemoryDescriptor values and supports the usual operations (indexing/iteration).

Methods

impl MemoryMap

unsafe fn from_raw(ptr: *mut MemoryDescriptor, memory_map_size: usize, descriptor_size: usize, descriptor_version: u32) -> Self

Constructs a memory map.

This takes ownership of the underlying buffer, and will deallocate it on drop.

This is a low-level constructor: you most likely want to use BootServices::memory_map() instead.

Panics

Panics if:

The pointer is null

The descriptor size is zero

The memory map size is not a multiple of the descriptor size

Safety

This is unsafe because the arguments must specify a valid memory map.

fn len(&self) -> usize

Returns the number of entries in the memory map.

fn iter(&self) -> MemoryMapIter

Returns an iterator over the memory map.

fn iter_mut(&mut self) -> MemoryMapMutIter

Returns a mutable iterator over the memory map.